// apps/api/migrations/snapshots/ecd49632b438334ad484f091a38019463e962a8fb9a47a451c7d9be7c56e8cc3/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Book: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly isbn: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly author: CodecTypes['pg/text@1']['output'];
      readonly publisher: CodecTypes['pg/text@1']['output'] | null;
      readonly publicationYear: CodecTypes['pg/int4@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly categoryId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly BookCopy: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly bookId: CodecTypes['pg/int4@1']['output'];
      readonly barcode: CodecTypes['pg/text@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly shelfLocation: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Category: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly LecturerProfile: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly lecturerNumber: CodecTypes['pg/text@1']['output'];
      readonly faculty: CodecTypes['pg/text@1']['output'] | null;
      readonly studyProgram: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Permission: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly code: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly Role: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly RolePermission: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly roleId: CodecTypes['pg/int4@1']['output'];
      readonly permissionId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly StudentProfile: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly npm: CodecTypes['pg/text@1']['output'];
      readonly faculty: CodecTypes['pg/text@1']['output'] | null;
      readonly studyProgram: CodecTypes['pg/text@1']['output'] | null;
      readonly enrollmentYear: CodecTypes['pg/int4@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly username: CodecTypes['pg/text@1']['output'] | null;
      readonly fullName: CodecTypes['pg/text@1']['output'];
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly phone: CodecTypes['pg/text@1']['output'] | null;
      readonly refreshTokenHash: CodecTypes['pg/text@1']['output'] | null;
      readonly isActive: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
    readonly UserRole: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly userId: CodecTypes['pg/int4@1']['output'];
      readonly roleId: CodecTypes['pg/int4@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Book: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly isbn: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly author: CodecTypes['pg/text@1']['input'];
      readonly publisher: CodecTypes['pg/text@1']['input'] | null;
      readonly publicationYear: CodecTypes['pg/int4@1']['input'] | null;
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly categoryId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly BookCopy: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly bookId: CodecTypes['pg/int4@1']['input'];
      readonly barcode: CodecTypes['pg/text@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly shelfLocation: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Category: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly LecturerProfile: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly lecturerNumber: CodecTypes['pg/text@1']['input'];
      readonly faculty: CodecTypes['pg/text@1']['input'] | null;
      readonly studyProgram: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Permission: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly code: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly Role: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly RolePermission: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly roleId: CodecTypes['pg/int4@1']['input'];
      readonly permissionId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly StudentProfile: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly npm: CodecTypes['pg/text@1']['input'];
      readonly faculty: CodecTypes['pg/text@1']['input'] | null;
      readonly studyProgram: CodecTypes['pg/text@1']['input'] | null;
      readonly enrollmentYear: CodecTypes['pg/int4@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly User: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly username: CodecTypes['pg/text@1']['input'] | null;
      readonly fullName: CodecTypes['pg/text@1']['input'];
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly phone: CodecTypes['pg/text@1']['input'] | null;
      readonly refreshTokenHash: CodecTypes['pg/text@1']['input'] | null;
      readonly isActive: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
    readonly UserRole: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly userId: CodecTypes['pg/int4@1']['input'];
      readonly roleId: CodecTypes['pg/int4@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;